Add DataQuery, DataFilter and moved some code around (#113)
* Add DataQuery, DataFilter and moved some code around

* add to changelog
